Apply SetStaffDefRedrawFlags and cleanup
