+++ include/vrv/staffdef.h
@@ -80,11 +80,6 @@ public:
     ///@}
 
     /**
-     * See Object::SetStaffDefRedrawFlags
-     */
-    int SetStaffDefRedrawFlags(FunctorParams *functorParams) override;
-
-    /**
      * See Object::PrepareDuration
      */
     int PrepareDuration(FunctorParams *functorParams) override;
